requirements.txt
# Copyright 2021-2025 ONDEWO GmbH
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.


######################################################
# ONDEWO Open Source Libraries
######################################################
-r requirements-ondewo-clients.txt
beautifulsoup4>=4.12.3
dataclasses>=0.6
######################################################
# External Libaries
######################################################
dataclasses-json>=0.5.4
docker>=7.1.0
fastapi>=0.115.6
grpcio~=1.66.1
httpx>=0.27.2
jsonschema>=4.23.0
langcodes>=3.4.1
numpy>=2.1.1
ondewo-logging>=3.3.3
pip-licenses
py~=1.11.0
pydantic>=2.8.2
pytz>=2023.2
regex>=2024.4.16
requests>=2.32.2
semantic-version>=2.10.0
setuptools>=67.1.0
starlette>=0.40.0
typing_extensions~=4.12.2
uvicorn>=0.30.4

######################################################
# ONDEWO Proprietary Libraries
######################################################
